fix: rename GamePage and HomePage
