--- src/routes/Game/index.js
@@ -1,16 +1,0 @@
-const GamePage = ({onChangePage}) => {
-
-  const handleClick = () => {
-    console.log('####: <Game Page />');
-    onChangePage && onChangePage('app');
-  }
-
-  return (
-    <div>
-      <h1>This is Game Page!!!</h1>
-      <button onClick={handleClick}>Start game</button>
-    </div>
-  )
-}
-
-export default GamePage;
